backend/requirements.txt


Flask
Flask-JWT-Extended
PyMySQL
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
Werkzeug
pandas
fyers-apiv3
psycopg2
selenium
webdriver-manager
requests
pandas
mysql-connector-python
schedule
